requirements.txt
asyncpg
python-dotenv
python-telegram-bot==20.3
sqlalchemy
